+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* [release] Mark required for release notes langs with `*`.
+
 ## 0.3.9- 2021-06-17
 
 * Auto `pub get` on `l10n` `extract` and `generate` commands.